fix test build

+++ src/crypto.c
@@ -1,7 +1,7 @@
 #include <pb.h>
 #include <crypto.h>
 
-extern const uint32_t no_of_keys;
+extern const uint8_t no_of_keys;
 extern const struct pb_key keys[];
 
 uint32_t pb_crypto_get_key(uint32_t key_index, struct pb_key **key)
